+++ src/components/pages/Main/index.tsx
@@ -1,28 +1,14 @@
 import React, { useState } from "react";
-import { Match, Player } from "../../../types";
+import { GameHistory, Match, OnBreakState, Player } from "../../../types";
 import GenerateMatchesButton from "../../atoms/GenerateMatchesButton";
 import CompleteMatchesButton from "../../atoms/CompleteMatchesButton";
-
-interface GameHistory {
-  [playerId: number]: number;
-}
-
-type OnBreakState = {
-  [key: number]: boolean;
-};
+import WaitingPlayers from "../../molecules/WaitingPlayers";
+import PlayerCard from "../../molecules/PlayerCard";
 
 interface PairHistory {
   [playerId: number]: {
     [partnerId: number]: number;
   };
-}
-
-interface NewComponentProps {
-  player: Player;
-  gameHistory: GameHistory;
-  winCount: number;
-  onBreakToggle: (playerId: number) => void;
-  isPlayerInMatch: (playerId: number) => boolean; // Add this line to the interface
 }
 
 function MainComponent() {
@@ -517,104 +503,5 @@ function MainComponent() {
     </div>
   );
 }
-
-const PlayerCard = ({
-  player,
-  gameHistory,
-  winCount,
-  onBreakToggle,
-  isPlayerInMatch,
-}: NewComponentProps) => {
-  const gamesPlayed = gameHistory[player.id] || 0;
-
-  return (
-    <div className="bg-white p-4 rounded shadow flex gap-4 items-center">
-      <div className="font-bold">選手 {player.id}</div>
-      <div>試合数: {gamesPlayed}</div>
-      <div>勝利数: {winCount}</div>
-
-      {!isPlayerInMatch(player.id) && (
-        <div className="ml-auto">
-          <button
-            className="bg-blue-500 hover:bg-blue-700 text-white font-bold py-1 px-4 rounded inline-block"
-            onClick={() => onBreakToggle(player.id)}
-          >
-            休憩
-          </button>
-        </div>
-      )}
-    </div>
-  );
-};
-
-type WaitingPlayersProps = {
-  players: Player[];
-  isPlayerInMatch: (playerId: number) => boolean;
-  onBreak: OnBreakState;
-  selectedPlayer: {
-    matchIndex: number;
-    team: number;
-    playerIndex: number;
-  } | null;
-  matches: Match[];
-  setMatches: (matches: Match[]) => void;
-  setSelectedPlayer: (
-    player: {
-      matchIndex: number;
-      team: number;
-      playerIndex: number;
-    } | null
-  ) => void;
-  gameHistory: GameHistory;
-  wins: GameHistory;
-};
-
-const WaitingPlayers = ({
-  players,
-  isPlayerInMatch,
-  onBreak,
-  selectedPlayer,
-  matches,
-  setMatches,
-  setSelectedPlayer,
-  gameHistory,
-  wins,
-}: WaitingPlayersProps) => {
-  return (
-    <div className="mt-8">
-      <h2 className="text-xl font-bold mb-4">待機中の選手</h2>
-      <div className="grid grid-cols-2 lg:grid-cols-4 gap-2 md:gap-4">
-        {players
-          .filter(
-            (player) => !isPlayerInMatch(player.id) && !onBreak[player.id]
-          )
-          .map((player) => (
-            <div
-              key={player.id}
-              className="bg-white p-4 rounded shadow cursor-pointer hover:bg-gray-100"
-              onClick={() => {
-                if (selectedPlayer) {
-                  const team = selectedPlayer.team === 1 ? "team1" : "team2";
-                  const newMatches = [...matches];
-                  const currentMatch = newMatches[selectedPlayer.matchIndex];
-                  currentMatch[team][selectedPlayer.playerIndex] = player;
-                  setMatches(newMatches);
-                  setSelectedPlayer(null);
-                }
-              }}
-            >
-              <div className="text-center">選手{player.id}</div>
-              <div className="text-center text-gray-500">
-                試合数: {gameHistory[player.id] || 0}
-              </div>
-              <div className="text-center text-blue-500">
-                勝利数: {wins[player.id] || 0}
-              </div>
-            </div>
-          ))}
-      </div>
-    </div>
-  );
-};
 
 export default MainComponent;
